we're tracking marbles 3d y level now
long explanation:
if marble goes under a bridge, bridge needs to be drawn on top of marble. if marble jaunts over bridge, bridge needs to be drawn under marble. we do not have the luxury to place events in the 3d spaces that tell the 2d space to reorganize their layering, so instead im checking the marble's 3d global y position. not the 2d y position the camera sees, that one is specific to the camera. this new one is specific to the 3d world the marble inhabits. it's okay if that doesnt make sense because im typing this hyped on hyperfocus energy THE IMPORTANT PART is im gonna need to keep track of what (3d) y levels mark the ball being below an object. this wont work for everything obvs but it's an easy way to help with layering. might need to add some for x and z but it depends on the type of object basically